+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # v0.3
 
 - Add `get_project_info` to access new end-point giving project info.
+- Print a warning when setting an attribute name with non-lowercase characters
+- Crash if trying to upload an attribute name with special characters
 
 # v0.2